==> 02_DC.ipynb <==
# DOC
## LAST UPDATE: 2022-02-18
## IPDC PROCEDURE
## STEP 3 & 4 (D & C): DECLUSTER AND CLASSIFY GAPS

input: 

pickle files from step 00_import 
* "./data/pickle/H.gpickle"
* "./data/pickle/h.pickle"
* "./data/pickle/ebc.pickle" 

csv file from 02_DC *AFTER ADDING MANUAL CLASSIFICATION OF GAPS*
* 

output:

pickle file of declustered gaps 
* "./data/pickle/gaps_declustered.pickle"

folium map of declustered gaps and csv file for modification (manual analysis)
* "./analysis/gaps_declustered_table.csv"
* "./analysis/gaps_declustered_plot.html"

# SETUP

# import packages
%run -i packages.py

# import parameters for plotting
%run -i parameters_plot.py

# IMPORT OBJECTS FROM PREVIOUS STEPS

H = nx.read_gpickle("./data/pickle/H.gpickle")
h = ig.Graph.Read_Pickle("./data/pickle/h.pickle")

ebc = pd.read_pickle("./data/pickle/ebc.pickle")
mygaps = pd.read_pickle("./data/pickle/mygaps.pickle")

# NX attribute dicts (for plotting)
ced = nx.get_edge_attributes(H, "coord") # coordinates of edges dictionary ced
ted = nx.get_edge_attributes(H, "category_edge") # type of edges dictionary ted
tnd = nx.get_node_attributes(H, "category_node") # type of nodes dictionary tnd
cnd = nx.get_node_attributes(H, "coord") # coordinates of nodes dictionary cnd

# make subfolder for storing output
if not os.path.exists("./analysis"):
    os.mkdir("./analysis")

# RUN DECLUSTERING HEURISTIC

# set cut-off value for benefit metric, exclude gaps below
B_cutoff = 15300
mygaps = mygaps[mygaps["B"]>=B_cutoff].reset_index(drop = True)

# ADDING EBC VALUES / RANKING METRICS TO EACH EDGE in network "f" (which will be used for declustering)
f = h.copy()
for i in range(len(ebc)):
    f.es[i]["ebc"] = ebc.loc[i, "ebc_lambda"]
    

### make a subgraph of f that contains only overlapping gaps

my_edges = list(set([item for sublist in mygaps["path"] for item in sublist]))
c = f.copy()
c = c.subgraph_edges(edges = my_edges, 
                         delete_vertices=True)

cl = c.decompose() ### cl contains disconnected components 

gapranks = []
gapcoords = []

gapranks = []
gapcoords = []
gapedgeids = []

# looping through disconnected components
for comp in range(len(cl)):
    
    mc = cl[comp].copy() # mc: my component (current)

    #### decluster component:

    while len(mc.es) > 0:

        nodestack = [node.index for node in mc.vs() if mc.vs[node.index]["category_node"]=="multi" and mc.degree(node.index)!=2]
        nodecomb = [comb for comb in itertools.combinations(nodestack, 2)] ### all possible OD combis on that cluster
        sp = [] ### list of shortest paths between d1 multi nodes on that cluster:
        for mycomb in nodecomb:

            gsp = mc.get_shortest_paths(mycomb[0], 
                                                mycomb[1],
                                                weights = "length", 
                                                mode = "out", 
                                                output = "epath")[0]
            if gsp:
                sp.append(gsp)

        ### compute metrics for each path:
        if not sp:
            break

        lens = []
        cycs = []
        
        for p in sp:
            lens += [np.sum([mc.es[e]["length"] for e in p])]
            cycs += [np.sum([mc.es[e]["length"]*mc.es[e]["ebc"] for e in p])]
            
        norms = list(np.array(cycs)/np.array(lens))
        maxpath = sp[norms.index(max(norms))]
        gapranks.append(np.round(max(norms), 2))
        
        gapcoord_current = []
        edgeids_current = []
        
        
        for e in maxpath:
            edge_id = tuple(sorted(literal_eval(mc.es[e]["edge_id"])))
            edge_coords = [(c[1], c[0]) for c in ced[tuple(sorted(edge_id))].coords]
            gapcoord_current.append(edge_coords)
            edgeids_current.append(edge_id)
            
        mc.delete_edges(maxpath)
        gapcoords.append(gapcoord_current)
        gapedgeids.append(edgeids_current)


gap_dec = pd.DataFrame({"rank": gapranks, "coord": gapcoords, "id": gapedgeids})
gap_dec = gap_dec.sort_values(by = "rank", ascending = False).reset_index(drop = True)
gap_dec = gap_dec[gap_dec["rank"]>=B_cutoff].copy().reset_index(drop = True)

gap_dec.to_pickle("./data/pickle/gaps_declustered.pickle")

### export to csv to start manual analysis
gap_dec["address"] = None
gap_dec["class"] = None
gap_dec["comments"] = None
gap_dec[["rank", "address", "class", "comments"]].to_csv(path_or_buf = "./analysis/gaps_declustered_table.csv", 
                                                         sep = ";", 
                                                         index = True, 
                                                         encoding = "utf-8", 
                                                         decimal = ",")

# MAKE FOLIUM MAP OF DECLUSTERED GAPS

### SET UP MAP WITH FOLIUM

# make map object

# coopenhagen coordinates
mycity_coord = [55.6761, 12.5683]

m = folium.Map(location = mycity_coord, zoom_start = 13, tiles = None) 

# add tile layers to m
for key in ["osm", "st_lite_op3", "white_background", "Google Satellite"]:
    basemaps[key].add_to(m)
    
##### plot car, onlybike, and multinetworks #####
snw = folium.FeatureGroup(name = "All streets", show = True)
bnw = folium.FeatureGroup(name = "Bikeable", show = True) # bikeable edges 

sloc = []
bloc = []

for edge in H.edges:
    myloc = [(c[1], c[0]) for c in ced[tuple(sorted(edge))].coords]
    sloc.append(myloc)
    if ted[edge] != "car":
        bloc.append(myloc)
        
snw_line = folium.PolyLine(locations = sloc, weight = 2, color = "#dadada").add_to(snw)
bnw_line = folium.PolyLine(locations = bloc, weight = 3, color = "black").add_to(bnw)

snw.add_to(m)
bnw.add_to(m)


#### ADD TO MAP 


gaps_fg = folium.FeatureGroup("Declustered gaps", show = True)

for locs in gap_dec["coord"]:
    myline = folium.PolyLine(locations = locs,
                                weight = 6,
                                color = "black").add_to(gaps_fg) # black border
    # + randomly colored gaps (not classified yet):
    myline = folium.PolyLine(locations = locs,
                                weight = 4,
                                color = random.choice(["#33FFDA", "#0AB023", "#B00A60", "#0A3CB0"])).add_to(gaps_fg)

gaps_fg.add_to(m)


my_fg_nr = folium.FeatureGroup("Gap numbers", show = False)

# Add pop-ups with gap number

for i in range(len(gap_dec)):
    
    my_marker = folium.Marker(location = gap_dec.loc[i]["coord"][0][0], 
                                    popup = "Gap " + str(i)
                             ).add_to(my_fg_nr)
                  
my_fg_nr.add_to(m)

### ADD LAYER CONTROL AND SAVE / DISPLAY FOLIUM MAP
folium.LayerControl().add_to(m)  

m.save("./analysis/gaps_declustered_plot.html")

## NEXT STEP IS MANUAL CLASSIFICATION:
* use "./analysis/gaps_declustered_table.csv" and "./analysis/gaps_declustered_plot.html" to manually classify gaps
* column "class": enter gap class (e.g. "BR" for bridge, "ST" for street, etc.)
* if gap is not confirmed (data issue/error): leave "class" cell for that row empty
* columns "address" and "comments": optional
* save edited csv table as "./analysis/gaps_classified_table.csv"
* table is imported in the next step (notebook 03_plot) for visualization of results

# import manually classified gaps
gaps_class = pd.read_csv("./analysis/gaps_classified_table.csv", 
                         index_col = 0,
                         sep = ";",  
                         decimal = ",",
                        encoding = "utf-8")# , sep = ";", decimal = ",", index_col = 0)

# import gap info (coords, edge ids, rank)
gaps_info = pd.read_pickle("./data/pickle/gaps_declustered.pickle")

# merge gaps_class and gap_info to get all data on classified gaps
gaps = pd.merge(gaps_class, gaps_info, how = "outer", indicator = False, on = "rank")

# drop unclassified gaps (that were not confirmed as gaps)
gaps_final = gaps.drop(gaps[gaps["class"].isnull()].index).reset_index(drop=True)
gaps_final["gapnumber"] = gaps_final.index + 1
gaps_final.to_pickle("./data/pickle/gaps_final.pickle") # save final gaps to pickle

### END OF 02_DC notebook
